requirements.txt
anthropic>=0.40.0
claude-agent-sdk
Pillow>=10.0.0
google-generativeai>=0.8.0
requests>=2.31.0
python-dotenv>=1.0.0
